Adding auth to user jobs

+++ src/controllers/jobController.js
@@ -3,6 +3,7 @@ const { v4: uuidv4 } = require('uuid');
 
 const formidable = require('formidable-serverless');
 
+const firebaseApp = require('../config/firebaseApp');
 const firebaseAdmin = require('../config/firebaseAdmin');
 const JobModel = require('../models/jobModel');
 
@@ -17,93 +18,102 @@ const CloudStorage = new Storage({
 
 const addJob = async (req, res) => {
     try {
-        const form = new formidable.IncomingForm({ multiples: true });
+        const user = firebaseApp.auth().currentUser;
 
-        // Default implementation
-        form.parse(req, async (error, fields, files) => {
-            // Create validation of the fields and files
-            if (!fields.title || !fields.companyName || !fields.location || !fields.email || !fields.jobType || !fields.requiredSkills || !fields.jobDescription || !files.companyProfileImage) {
-                return res.status(400).json({
-                    message: 'Please Fill All Required Input Fields',
-                    status: 400
-                });
-            }
+        if (user && req.user.uid) {
+            const form = new formidable.IncomingForm({ multiples: true });
 
-            const id = uuidv4();
+            // Default implementation
+            form.parse(req, async (error, fields, files) => {
+                // Create validation of the fields and files
+                if (!fields.title || !fields.companyName || !fields.location || !fields.email || !fields.jobType || !fields.requiredSkills || !fields.jobDescription || !files.companyProfileImage) {
+                    return res.status(400).json({
+                        message: 'Please Fill All Required Input Fields',
+                        status: 400
+                    });
+                }
 
-            const bucketName = process.env.GOOGLE_CLOUD_STORAGE_BUCKET_NAME;
+                const id = uuidv4();
 
-            const storagePublicURL = `https://storage.googleapis.com/${bucketName}.appspot.com/`;
+                const bucketName = process.env.GOOGLE_CLOUD_STORAGE_BUCKET_NAME;
 
-            // const storagePublicURL = `https://firebasestorage.googleapis.com/v0/b/${bucketName}.appspot.com/o/`;
+                const storagePublicURL = `https://storage.googleapis.com/${bucketName}.appspot.com/`;
 
-            // The variable should be match with the name of the key field
-            const companyProfileImage = files.companyProfileImage;
+                // const storagePublicURL = `https://firebasestorage.googleapis.com/v0/b/${bucketName}.appspot.com/o/`;
 
-            // URL of the uploaded image
-            let imageURL;
+                // The variable should be match with the name of the key field
+                const companyProfileImage = files.companyProfileImage;
 
-            const jobID = JobsCollection.doc().id;
+                // URL of the uploaded image
+                let imageURL;
 
-            if (error) {
-                return res.status(400).json({
-                    message: 'There Was an Error Parsing The Files',
-                    status: 400,
-                    error: error.message
-                });
-            }
+                const jobID = JobsCollection.doc().id;
 
-            const bucket = CloudStorage.bucket(`gs://${bucketName}.appspot.com`);
+                if (error) {
+                    return res.status(400).json({
+                        message: 'There Was an Error Parsing The Files',
+                        status: 400,
+                        error: error.message
+                    });
+                }
 
-            if (companyProfileImage.size === 0) {
-                res.status(404).send({
-                    message: 'No Image Found',
-                    status: 404
-                });
-            } else {
-                const imageResponse = await bucket.upload(companyProfileImage.path, {
-                    destination: `${process.env.JOBS_COLLECTION}/${jobID}/${companyProfileImage.name}`,
-                    resumable: true,
-                    metadata: {
+                const bucket = CloudStorage.bucket(`gs://${bucketName}.appspot.com`);
+
+                if (companyProfileImage.size === 0) {
+                    res.status(404).send({
+                        message: 'No Image Found',
+                        status: 404
+                    });
+                } else {
+                    const imageResponse = await bucket.upload(companyProfileImage.path, {
+                        destination: `${process.env.JOBS_COLLECTION}/${user.uid || req.user.uid}/${jobID}/${companyProfileImage.name}`,
+                        resumable: true,
                         metadata: {
-                            firebaseStorageDownloadTokens: id
+                            metadata: {
+                                firebaseStorageDownloadTokens: id
+                            }
                         }
-                    }
-                });
+                    });
 
-                // Profile image url
-                // imageURL = `${storagePublicURL + encodeURIComponent(imageResponse[0].name)}?alt=media&token=${id}`;
+                    // Profile image url
+                    // imageURL = `${storagePublicURL + encodeURIComponent(imageResponse[0].name)}?alt=media&token=${id}`;
 
-                imageURL = storagePublicURL + imageResponse[0].name;
-            }
+                    imageURL = storagePublicURL + imageResponse[0].name;
+                }
 
-            const date = new Date();
+                const date = new Date();
 
-            const getDateAndTime = date.toLocaleDateString() + '|' + date.toLocaleTimeString();
+                const getDateAndTime = date.toLocaleDateString() + '|' + date.toLocaleTimeString();
 
-            // Object to send to the database
-            const jobData = {
-                id: jobID,
-                title: fields.title,
-                companyName: fields.companyName,
-                location: fields.location,
-                email: fields.email,
-                jobType: fields.jobType,
-                requiredSkills: fields.requiredSkills,
-                jobDescription: fields.jobDescription,
-                companyProfileImage: companyProfileImage.size === 0 ? '' : imageURL,
-                createdAt: getDateAndTime
-            };
+                // Object to send to the database
+                const jobData = {
+                    id: jobID,
+                    title: fields.title,
+                    companyName: fields.companyName,
+                    location: fields.location,
+                    email: fields.email,
+                    jobType: fields.jobType,
+                    requiredSkills: fields.requiredSkills,
+                    jobDescription: fields.jobDescription,
+                    companyProfileImage: companyProfileImage.size === 0 ? '' : imageURL,
+                    createdAt: getDateAndTime
+                };
 
-            // Added to the firestore collection
-            await JobsCollection.doc(jobID).set(jobData, { merge: true }).then(() => {
-                res.status(201).send({
-                    message: 'Successfully Added Job',
-                    status: 201,
-                    data: jobData
+                // Added to the firestore collection
+                await JobsCollection.doc(jobID).set(jobData, { merge: true }).then(() => {
+                    res.status(201).send({
+                        message: 'Successfully Added Job',
+                        status: 201,
+                        data: jobData
+                    });
                 });
             });
-        });
+        } else {
+            res.status(403).send({
+                message: 'User is Not Sign In',
+                status: 403
+            });
+        }
     } catch (error) {
         res.status(400).send({
             message: 'Something Went Wrong to Added Job',
@@ -125,6 +135,18 @@ const getAllJobs = async (req, res) => {
     }
 };
 
+const getAllUserJobs = async (req, res) => {
+    try {
+        JobModel.getAllUserJobs(req, res, JobsCollection);
+    } catch (error) {
+        res.status(400).send({
+            message: 'Something Went Wrong to Display All User Jobs',
+            status: 400,
+            error: error.message
+        });
+    }
+};
+
 const getJobDetail = async (req, res) => {
     try {
         JobModel.getJobDetail(req, res, JobsCollection);
@@ -140,95 +162,104 @@ const getJobDetail = async (req, res) => {
 // Update job profile by id (Only for Testing)
 const updateJob = async (req, res) => {
     try {
-        const jobID = req.params.id;
-        const job = await JobsCollection.doc(jobID).get();
+        const user = firebaseApp.auth().currentUser;
 
-        const form = new formidable.IncomingForm({ multiples: true });
+        if (user && req.user.uid) {
+            const jobID = req.params.id;
+            const job = await JobsCollection.doc(jobID).get();
 
-        if (!job.exists) {
-            res.status(404).send({
-                message: 'Job is Not Found',
-                status: 404
-            });
-        } else {
-            // Default implementation
-            form.parse(req, async (error, fields, files) => {
-                // Create validation of the fields and files
-                if (!fields.title || !fields.companyName || !fields.location || !fields.email || !fields.jobType || !fields.requiredSkills || !fields.jobDescription || !files.companyProfileImage) {
-                    return res.status(400).json({
-                        message: 'Please Fill All Required Input Fields',
-                        status: 400
-                    });
-                }
+            const form = new formidable.IncomingForm({ multiples: true });
 
-                const bucketName = process.env.GOOGLE_CLOUD_STORAGE_BUCKET_NAME;
+            if (!job.exists) {
+                res.status(404).send({
+                    message: 'Job is Not Found',
+                    status: 404
+                });
+            } else {
+                // Default implementation
+                form.parse(req, async (error, fields, files) => {
+                    // Create validation of the fields and files
+                    if (!fields.title || !fields.companyName || !fields.location || !fields.email || !fields.jobType || !fields.requiredSkills || !fields.jobDescription || !files.companyProfileImage) {
+                        return res.status(400).json({
+                            message: 'Please Fill All Required Input Fields',
+                            status: 400
+                        });
+                    }
 
-                const storagePublicURL = `https://storage.googleapis.com/${bucketName}.appspot.com/`;
+                    const bucketName = process.env.GOOGLE_CLOUD_STORAGE_BUCKET_NAME;
 
-                // const storagePublicURL = `https://firebasestorage.googleapis.com/v0/b/${bucketName}.appspot.com/o/`;
+                    const storagePublicURL = `https://storage.googleapis.com/${bucketName}.appspot.com/`;
 
-                const companyProfileImage = files.companyProfileImage;
+                    // const storagePublicURL = `https://firebasestorage.googleapis.com/v0/b/${bucketName}.appspot.com/o/`;
 
-                // URL of the uploaded image
-                let imageURL;
+                    const companyProfileImage = files.companyProfileImage;
 
-                if (error) {
-                    return res.status(400).json({
-                        message: 'There Was an Error Parsing The Files',
-                        status: 400,
-                        error: error.message
-                    });
-                }
+                    // URL of the uploaded image
+                    let imageURL;
 
-                const bucket = CloudStorage.bucket(`gs://${bucketName}.appspot.com`);
+                    if (error) {
+                        return res.status(400).json({
+                            message: 'There Was an Error Parsing The Files',
+                            status: 400,
+                            error: error.message
+                        });
+                    }
 
-                if (companyProfileImage.size === 0) {
-                    res.status(404).send({
-                        message: 'No Image Found',
-                        status: 404
-                    });
-                } else {
-                    const imageResponse = await bucket.upload(companyProfileImage.path, {
-                        destination: `${process.env.JOBS_COLLECTION}/${jobID}/${companyProfileImage.name}`,
-                        resumable: true,
-                        metadata: {
+                    const bucket = CloudStorage.bucket(`gs://${bucketName}.appspot.com`);
+
+                    if (companyProfileImage.size === 0) {
+                        res.status(404).send({
+                            message: 'No Image Found',
+                            status: 404
+                        });
+                    } else {
+                        const imageResponse = await bucket.upload(companyProfileImage.path, {
+                            destination: `${process.env.JOBS_COLLECTION}/${jobID}/${companyProfileImage.name}`,
+                            resumable: true,
                             metadata: {
-                                firebaseStorageDownloadTokens: jobID
+                                metadata: {
+                                    firebaseStorageDownloadTokens: jobID
+                                }
                             }
-                        }
-                    });
+                        });
 
-                    // Profile image url
-                    // imageURL = `${storagePublicURL + encodeURIComponent(imageResponse[0].name)}?alt=media&token=${uuid}`;
+                        // Profile image url
+                        // imageURL = `${storagePublicURL + encodeURIComponent(imageResponse[0].name)}?alt=media&token=${uuid}`;
 
-                    imageURL = storagePublicURL + imageResponse[0].name;
-                }
+                        imageURL = storagePublicURL + imageResponse[0].name;
+                    }
 
-                const date = new Date();
+                    const date = new Date();
 
-                const getDateAndTime = date.toLocaleDateString() + '|' + date.toLocaleTimeString();
+                    const getDateAndTime = date.toLocaleDateString() + '|' + date.toLocaleTimeString();
 
-                // Object to send to the database
-                const jobData = {
-                    title: fields.title,
-                    companyName: fields.companyName,
-                    location: fields.location,
-                    email: fields.email,
-                    jobType: fields.jobType,
-                    requiredSkills: fields.requiredSkills,
-                    jobDescription: fields.jobDescription,
-                    companyProfileImage: companyProfileImage.size === 0 ? '' : imageURL,
-                    updatedAt: getDateAndTime
-                };
+                    // Object to send to the database
+                    const jobData = {
+                        title: fields.title,
+                        companyName: fields.companyName,
+                        location: fields.location,
+                        email: fields.email,
+                        jobType: fields.jobType,
+                        requiredSkills: fields.requiredSkills,
+                        jobDescription: fields.jobDescription,
+                        companyProfileImage: companyProfileImage.size === 0 ? '' : imageURL,
+                        updatedAt: getDateAndTime
+                    };
 
-                // Update to the firestore collection
-                await JobsCollection.doc(jobID).update(jobData, { merge: true }).then(() => {
-                    res.status(202).send({
-                        message: 'Update Job Successfully',
-                        status: 202,
-                        data: jobData
+                    // Update to the firestore collection
+                    await JobsCollection.doc(jobID).update(jobData, { merge: true }).then(() => {
+                        res.status(202).send({
+                            message: 'Update Job Successfully',
+                            status: 202,
+                            data: jobData
+                        });
                     });
                 });
+            }
+        } else {
+            res.status(403).send({
+                message: 'User is Not Sign In',
+                status: 403
             });
         }
     } catch (error) {
@@ -265,22 +296,31 @@ const deleteJobProfileStorage = async (uid) => {
 // Delete job profile by id (Only for Testing)
 const deleteJob = async (req, res) => {
     try {
-        const jobID = req.params.id;
-        const job = await JobsCollection.doc(jobID).get();
+        const user = firebaseApp.auth().currentUser;
 
-        if (!job.exists) {
-            res.status(404).send({
-                message: 'Job is Not Found',
-                status: 404
-            });
+        if (user || req.user.uid) {
+            const jobID = req.params.id;
+            const job = await JobsCollection.doc(jobID).get();
+
+            if (!job.exists) {
+                res.status(404).send({
+                    message: 'Job is Not Found',
+                    status: 404
+                });
+            } else {
+                await JobsCollection.doc(req.params.id).delete();
+
+                deleteJobProfileStorage(jobID);
+
+                res.status(202).send({
+                    message: 'Delete Job Successfully',
+                    status: 202
+                });
+            }
         } else {
-            await JobsCollection.doc(req.params.id).delete();
-
-            deleteJobProfileStorage(jobID);
-
-            res.status(202).send({
-                message: 'Delete Job Successfully',
-                status: 202
+            res.status(403).send({
+                message: 'User is Not Sign In',
+                status: 403
             });
         }
     } catch (error) {
@@ -292,4 +332,4 @@ const deleteJob = async (req, res) => {
     }
 };
 
-module.exports = { addJob, getAllJobs, getJobDetail, updateJob, deleteJob };
+module.exports = { addJob, getAllJobs, getAllUserJobs, getJobDetail, updateJob, deleteJob };
